split server and app for testing
