+++ server/package.json
@@ -4,14 +4,24 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
+  },
+  "jest": {
+    "transform": {
+      "^.+\\.[t|j]sx?$": "babel-jest"
+    }
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "apollo-server-express": "^3.5.0",
+    "dot-env": "^0.0.1",
     "express": "^4.17.2",
     "mongoose": "^6.1.3"
+  },
+  "devDependencies": {
+    "babel-jest": "^27.4.6",
+    "jest": "^27.4.7"
   }
 }
